% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNAGenes.updated.R
\name{miRNAGenes.updated}
\alias{miRNAGenes.updated}
\title{Get microRNA-target Interactions}
\usage{
miRNAGenes.updated(
  DE.miRNA,
  DE.target = NULL,
  path = "/bicoh/MARGenomics/annotationData/miRNAIntegration/"
)
}
\arguments{
\item{DE.miRNA}{Character string or character vector of miRNA(s)}

\item{DE.target}{Character string or character vector of mRNA(s)}

\item{path}{Character string with the path to where the database files are stored}
}
\value{
Returns a list with four elements: 1) list of  miRNAs each conataining 
a vector with all identified targets in DE.target; 2) dataframe with results from 
miRDB, 3) mirTarBase and4) TargetScan
}
\description{
Retrieve predicted and validated miRNA-target interactions using TargetScan,
 miRDB and mirTarBase
}
\author{
Julia Perera Bel <jperera@imim.es>
}
