% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/miRNAGenes.R
\name{miRNAGenes}
\alias{miRNAGenes}
\title{Get microRNA-target Interactions}
\usage{
miRNAGenes(DE.miRNA, DE.target)
}
\arguments{
\item{DE.miRNA}{Character string or character vector of miRNA(s)}

\item{DE.target}{Character string or character vector of mRNA(s)}
}
\value{
Returns a list of miRNAs, each containing a vector with all 
identified targets in DE.target
}
\description{
Retrieve predicted and validated miRNA-target interactions using Multimir,
miRNAtap and SpideRmiR R packages
}
\details{
Requires several packages: multiMiR, miRNAtap, miRNAtap.db and 
SpidermiR. Not recomended to use this function as the database versions 
used in these packages are not up to date
}
\author{
Julia Perera Bel <jperera@imim.es>
}
